Merge mutated tests which use the same mutation
When we mutate unit tests, for each specific mutation (e.g. replacing a
function with an iftrue'd version) we want *one of* the unit tests to
fail because of the mutation. Having all of them fail is nice but too
strong a requirement.

To do this, we now combine all the unit tests (for each search/replace
pair) by making them non-#[test] functions scoped to one large #[test]
function which calls all its children. The result is that it will fail
if any of its children fail, giving the desired behaviour.